[WIP!!!][BUGFIX] Constant FLUIDCONTENT_TEMPFILE is not always defined
Constant FLUIDCONTENT_TEMPFILE is defined in ext_tables.php which is
meant to point to a file to cache Page TSConfig.

However, this leads to a problem as of TYPO3 6.1 which has a refactored
caching mechanism. After the global TYPO3 cache has been set (processing
of ext_tables.php and ext_localconf.php among other), the constants
is never defined anymore. As a result Fluid Content is not able
to cache Page TSConfig anymore impacting the overall performance.