[BUGFIX] Use correct content element icon path
Icons in the new content element wizard are broken if
the TYPO3 root directory is a symlink.

realpath() was used to fix a double-slash issue in the icon path,
but this breaks PATH_site detection.
Instead of relying on realpath to fix it, we correctly remove the leading
slash now, keeping the original path.

Without the fix, your browser will try to load icons with their full disk path
like
http://localhost/var/www/site/typo3conf/ext/foo/Resources/Public/Icons/Content/Foo.svg