[BUGFIX] Fix copy translation workflow
This fix adds a postProcessRecord hook, that checks,
if a fluidcontent record's sys_language_uid is updated
and applies this new language to all child elements
of the updated one. Without this, every content
element inside the root content element needs to be
manually changed to the target language after a
"Copy" translation in the backend.